fix pr comments

+++ linkis-computation-governance/linkis-manager/linkis-application-manager/src/main/scala/org/apache/linkis/manager/rm/utils/AcrossClusterRulesJudgeUtils.scala
@@ -26,18 +26,19 @@ object AcrossClusterRulesJudgeUtils extends Logging {
       leftResource: YarnResource,
       maxResource: YarnResource,
       leftCPUThreshold: Int,
-      leftMemoryThreshold: Long,
+      leftMemoryThreshold: Int,
       UsedCPUPercentageThreshold: Double,
       UsedMemoryPercentageThreshold: Double
   ): Boolean = {
     if (leftResource != null && maxResource != null) {
       val leftQueueMemory = leftResource.queueMemory / Math.pow(1024, 3).toLong
-
       logger.info(
         s"leftResource.queueCores: ${leftResource.queueCores}, leftCPUThreshold: $leftCPUThreshold"
       )
-      logger.info(s"leftQueueMemory: $leftQueueMemory, leftMemoryThreshold: $leftMemoryThreshold")
-      if (leftResource.queueCores > leftCPUThreshold && leftQueueMemory > leftMemoryThreshold) { // 目标集群的剩余资源高于阈值，资源充足
+      logger.info(
+        s"leftQueueMemory: $leftQueueMemory, leftMemoryThreshold: $leftMemoryThreshold"
+      )
+      if (leftResource.queueCores > leftCPUThreshold && leftQueueMemory > leftMemoryThreshold) {
         val maxQueueMemory = maxResource.queueMemory / Math.pow(1024, 3).toLong
         val usedCPUPercentage =
           1 - leftResource.queueCores.toDouble / maxResource.queueCores.toDouble
@@ -52,7 +53,7 @@ object AcrossClusterRulesJudgeUtils extends Logging {
         )
         if (
             usedCPUPercentage < UsedCPUPercentageThreshold && usedMemoryPercentage < UsedMemoryPercentageThreshold
-        ) { // 目标集群的资源使用率小于阈值，资源充足
+        ) {
           return true
         }
       }
